requirements.txt
click==8.0.3
colorama==0.4.4
deap==1.3.1
numpy==1.21.2
PySide6==6.2.0
PyYAML==6.0
shiboken6==6.2.0
